FSType update fixes

+++ Install/INCLUDE/Ptab.h
@@ -112,15 +112,15 @@ class CPartList {
 		int CanActivate(int Index);
 		void SetAllowActiveHD(int Status);
 		void SetActive(int Index);
-		void SetFsType(int Index, int FsType);
+		void SetFsType(int Index, unsigned short FsType);
 		int UpgradeXoslBootItem(const TPartition *Partition,unsigned char MbrHDSector0);
 		int Retain(const char *DosFileName,unsigned short FileSize,const TPartition *Partition);
 		void UpdateFSType(int Index, unsigned short FSType, unsigned char MbrHDSector0);
 		int GetGPTIndex(uuid_t GPTType);
 		unsigned short GetGPTShortType(uuid_t GPTType);
-		uuid_t* GetGPTType(int FSType);
+		uuid_t* GetGPTType(unsigned short FSType);
 		uint16_t GetGptMBRType(int gpt_index);
-		char* GetGPTName(int FSType);
+		char* GetGPTName(unsigned short FSType);
 	public:
 		typedef struct {
 			int FSID;
